pykupid/requirements.txt
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
Pillow==9.4.0
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
